requirements.txt
pygame-ce
Pillow
moderngl
